+++ app/templates/displayFilesMacro.html
@@ -1,17 +1,18 @@
-{% macro displayFiles(filepaths, titleName, deleteLink, databaseId, prefixToRemove="")  %}
+{% macro displayFiles(filePaths, titleName, deleteLink, databaseId)  %}
+
 <table>
     <tr>
         <td class="py-2"><strong>Display</strong></td>
         <td class="py-2 px-3"><strong>{{ titleName }}</strong></td>
     </tr>
-    {% for key, value in filepaths.items() %}
+    {% for key, value in filePaths.items() %}
         <tr id="attachment_{{ value[1] }}" class="attachmentRow_{{ key }}" data-id="{{ key }}">
             <td class="py-2 px-3">
                 {% if value[-1].isDisplayed %}
                     <input type="radio" class="attachmentCheck" data-id="{{ value[1] }}" checked>
                 
                 {% else %}
-                {% if not (value[0].endswith('.png') or value[0].endswith('.jpg') or value[0].endswith('.svg') or value[0].endswith('.jpeg') or value[0].endswith('.gif')) %}
+                    {% if not (value[0].endswith('.png') or value[0].endswith('.jpg') or value[0].endswith('.svg') or value[0].endswith('.jpeg') or value[0].endswith('.gif')) %}
                         -
                     {% else %}
                         <input type="radio" class="attachmentCheck" data-id="{{ value[1] }}" >
@@ -19,9 +20,10 @@
                 {% endif %}
             </td>
             <td class="py-2 px-3" >
-                {% set filename = key.removeprefix(prefixToRemove) %}
-                {% set shortname = filename[:8] + "..." + filename[-10:] if filename|length > 25 else filename %}
-                <a class="mr-5 fileName" data-filename='{{ filename }}' href="{{ value[0] }}" target="_blank" data-toggle="tooltip" data-placement="top" title="{{filename}}" aria-labelledby="{{filename}}">{{ shortname }}</a>
+                {% set idx = key.index("/") %}
+                {% set fileName = key[idx+1:] %}
+                {% set shortName = fileName[:8] + "..." + fileName[-10:] if fileName|length > 25 else fileName %}
+                <a class="mr-5 fileName" data-filename='{{ fileName }}' href="{{ value[0] }}" target="_blank" data-toggle="tooltip" data-placement="top" title="{{fileName}}" aria-labelledby="{{fileName}}">{{ shortName }}</a>
             </td>
             <td style="text-align:center">
                 <button
