+++ app/templates/eventView.html
@@ -150,9 +150,11 @@
                         <label class="mb-2" for= "eventAttachments"><strong>Event Attachments</strong></label>
                         <table>
                             {% for key, value in filepaths.items() %}
+                                {% set idx = key.index("/") %}
+                                {% set fileName = key[idx+1:] %}
                                 <tr id="attachment_{{value[1]}}" class="attachmentRow_{{key}}" data-id= "{{key}}">
                                     <td class="pr-5">
-                                        <a class="mr-5" href="{{value[0]|replace('#','%23')}}" target="_blank">{{key}}</a>
+                                        <a class="mr-5" href="{{value[0]|replace('#','%23')}}" target="_blank">{{fileName}}</a>
                                     </td>
                                 </tr>
                             {% endfor %}
