+++ app/templates/serviceLearning/slcQuestionnaire.html
@@ -9,6 +9,8 @@
 <p>Referring to the Criteria for Service-Learning Course Designation/ALE Requirement met through Service-Learning, please address the following in a
    one-page proposal narrative. You might aim for approximately 500 words in total. Attach your syllabus (or the one form the most recent time that the course was taught). If you are in the process of
    designing the syllabus, please indicate that.</p>
+
+
 <div class="pt-4">
   <label class="pb-3"><b>
     1. Indicate the specific course learning objective(s) that will be addressed by the service-learning component.
